feat(config-ui): rename the config to the connection
