+++ config-ui-lib/packages/connection/src/components/rate-limit/styled.ts
